Refactor forwarding and logs
