+++ aws/logs_monitoring/logs/helpers.py
@@ -1,3 +1,9 @@
+# Unless explicitly stated otherwise all files in this repository are licensed
+# under the Apache License Version 2.0.
+# This product includes software developed at Datadog (https://www.datadoghq.com/).
+# Copyright 2021 Datadog, Inc.
+
+
 import logging
 import re
 import gzip
